requirements.txt
einops==0.8.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
importlib_metadata==8.7.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
numpy==2.2.6
pandas==2.2.3
pytorch-lightning==2.5.1.post0
rl4co>=0.6.0
torch==2.8.0
tqdm==4.67.1